P0-3: Fix Money Fields Float → Decimal (12 fields, 6 models)
- Changed 12 Float money fields to Decimal across 6 models:
  Device.purchaseCost, MaintenanceRecord.cost,
  Plan.priceMonthly/priceAnnual, Invoice.amount/tax/total,
  Quotation.subtotal/taxRate/tax/total, QuotationItem.unitPrice
- Added Decimal.prototype.toJSON patch in db.ts for transparent
  number serialization in API JSON responses
- Fixed 5 API routes for server-side Decimal arithmetic:
  reports, revenue-forecast, maintenance-prediction,
  invoice PDF, quotations PATCH
- 15 non-money Float fields verified unchanged
- 38 new tests in tests/money-decimal-p0.test.ts
- Updated quotation-items test for Decimal type
- All 281 tests pass, 0 TS errors, 0 ESLint errors, build passes

+++ tests/quotation-items-p1.test.ts
@@ -34,7 +34,7 @@ describe('P1-9: Quotation Items Normalization', () => {
       expect(model).toMatch(/sortOrder\s+Int/);
       expect(model).toMatch(/description\s+String/);
       expect(model).toMatch(/quantity\s+Int/);
-      expect(model).toMatch(/unitPrice\s+Float/);
+      expect(model).toMatch(/unitPrice\s+Decimal/);
       
       // Relation
       expect(model).toMatch(/quotation\s+Quotation/);
